packages/auth-client/pyproject.toml
[project]
name = "auth-client"
version = "0.1.0"
description = "Python client library for the SignalForge Auth Service"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.28.0,<1.0.0",
    "pydantic>=2.8.0,<3.0.0",
    "python-jose[cryptography]>=3.3.0,<4.0.0",
    "fastapi>=0.115.0,<1.0.0",
]

[project.optional-dependencies]
test = [
    "pytest>=8.0.0,<9.0.0",
    "pytest-asyncio>=0.23.0,<1.0.0",
    "respx>=0.21.0,<1.0.0",
    "httpx>=0.28.0,<1.0.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = ["tests"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/auth_client"]

[tool.uv]
package = true
